split briefing terms, re-added some Eat It stuff
you might skip this commit if you don't like it.  I put back in the under-lava tunnels from Eat It, but not as large and without the platform - instead, you swim out.  Then I put in the first page of the civil war term.   Finally, I split the briefing into two parts, one you get to via the platform in the second room, the second part where you originally see it in the tower.