Third chapter platforms, sounds, bug fixes, &c.
I haven’t gone through *all* the platforms to make sure they run at a reasonable speed, but I’ve added sound objects to all the ones that looked like they needed them. I also added a 3x recharge powerup to the start of “S’pht’ia” as discussed on the Pfhorums, renamed “Flames of Icarus” to “Flight of Icarus” given the lack of lava in the new version (also as discussed on the Pfhorums), revised the sound objects in “Icarus” to reflect the new media (sewage now has sewage sound objects floating on it, and likewise for water), and fixed an apparently unintentional ceiling in “Icarus” that appears to have made it impossible to complete the level. There might’ve been a couple of other changes, but it was 2:10 am when I originally wrote this description, and I was pretty much brain-dead. (It’s 2:10 pm now, and I’m still pretty much brain-dead.)