Hysterical Womb updates from Aaeron
"Made some changes in anticipation of rearranging the lights to create a lightning effect, but haven’t yet gotten around to that. This still flickers the fog and turns tag 7 off & on (tag 7 is what will control the lightning). There’s only one small segment of the level that’s even outdoors, so it won’t affect much of the level, but this way at least the locations on this planet in the first half of this chapter will have some consistent aspects to their setting"