Update Hysterical-Womb.sceA
mistextured floors and aaron's changes: "I also rolled back the lightning, since this is not supposed to take place on the Pfhor planet; this also necessitated a change to the Lua file. We’ll also need to modify text resource 128 to remove the ambient sounds for rain on this level, which I’ll do at some later point.

I also added a flickering light to some of the goo in various parts of the level and put in the “transformer” sound in teleporter locations (except the under-the-goo teleporters), which I’m going to make a scenario-wide convention as time permits."