Update A-Friend-In-Need.sceA
mistextured sky and aaron's changes: "I’ve also set it so there’s a forcefield directly south of the chip insertion point, and inserting the chip turns it off. I’m still not sure if I like this better than how it was (the contrast between the two sky types looked neat, but it makes more story sense for there to be a forcefield there as well); I may revert the behaviour.

It may additionally be worth revising the terminals to make it more explicit where the player is supposed to teleport out, amongst other things. I’ll take a crack at that this weekend, perhaps.

It may actually be worth setting a forcefield to turn on when the platform with the formerly mistextured sky finishes rising, just to take it out of the pool of rain polygons. That would require a slight script revision, but shouldn’t be too complicated; I might take a crack at that this weekend as well."