s'pht'ia updates
terminal tweaks and aaron's changes: "Fixed the motion glitch that occurred when a player accessed the secret from the start position.

Also coded a potential fix to players not getting 3x shields when they revive in network games, but haven't been able to test it yet. We'll find out if it works during our next net game, presumably."