Update These-Caves-Can’t-Be-a-Natural-Formation.sceA
aaron's updates: "Fixed a lot of the remaining motion glitches. How to fix the glitch with line 2587 is beyond me, though. ¯(°_o)/¯

Note that I left a couple of specific motion glitches (lines 622 and 626) because fixing them would’ve probably made it impossible to do a particular grenade jump that can be used during a speedrun. I probably could’ve fixed 556 and 560 without such issues, but it didn’t occur to me at the time. I’ll get around to it later."